Removing setting -config argument as configuration is passed as an environmental variable now
