+++ tests/setting_properties/test_command_line_explorer.py
@@ -2,10 +2,10 @@ import os
 import shutil
 import sumstats.snp.loader as loader
 from tests.prep_tests import *
-from sumstats import controller as controller
+from sumstats import explorer as explorer
 
 
-class TestCommandLineController(object):
+class TestCommandLineExplorer(object):
 
     def setup_method(self):
         os.makedirs('./output/bysnp/1')
@@ -24,11 +24,6 @@ class TestCommandLineController(object):
     def teardown_method(self):
         shutil.rmtree('./output')
 
-    def test_controller(self):
-        parser = controller.argument_parser(['-path', './output', '-snp', 'rs185339560'])
-        assert parser.path == "./output"
-        assert parser.snp == "rs185339560"
-
-    def test_controller_config(self):
-        parser = controller.argument_parser(['-config', 'config/properties.json'])
-        assert parser.config == "config/properties.json"
+    def test_explorer(self):
+        parser = explorer.argument_parser(['-traits'])
+        assert parser.traits == True
